# backend/public/swagger.json
{
	"components": {
...
		"schemas": {
...
			"ResponseUserRegister": {
				"properties": {
					"is_success": {
						"type": "boolean"
					},
					"error_code": {
						"type": "number",
						"format": "double"
					},
					"error_msg": {
						"type": "string"
					},
					"user_uuid": {
						"type": "string"
					}
				},
				"required": [
					"is_success",
					"error_code",
					"error_msg",
					"user_uuid"
				],
				"type": "object",
				"additionalProperties": false
			},
			"RequestUserRegister": {
				"properties": {
					"request_hash": {
						"type": "string"
					},
					"timestamp": {
						"type": "number",
						"format": "double"
					},
					"username": {
						"type": "string"
					},
					"password_hash": {
						"type": "string"
					}
				},
				"required": [
					"request_hash",
					"timestamp",
					"username",
					"password_hash"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ResponseUserLogin": {
				"properties": {
					"is_success": {
						"type": "boolean"
					},
					"error_code": {
						"type": "number",
						"format": "double"
					},
					"error_msg": {
						"type": "string"
					},
					"session_hash": {
						"type": "string"
					},
					"user_uuid": {
						"type": "string"
					}
				},
				"required": [
					"is_success",
					"error_code",
					"error_msg",
					"session_hash",
					"user_uuid"
				],
				"type": "object",
				"additionalProperties": false
			},
			"RequestUserLogin": {
				"properties": {
					"request_hash": {
						"type": "string"
					},
					"timestamp": {
						"type": "number",
						"format": "double"
					},
					"username": {
						"type": "string"
					},
					"password_hash": {
						"type": "string"
					}
				},
				"required": [
					"request_hash",
					"timestamp",
					"username",
					"password_hash"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ResponseUserGetUsername": {
				"properties": {
					"is_success": {
						"type": "boolean"
					},
					"error_code": {
						"type": "number",
						"format": "double"
					},
					"error_msg": {
						"type": "string"
					},
					"username": {
						"type": "string"
					}
				},
				"required": [
					"is_success",
					"error_code",
					"error_msg",
					"username"
				],
				"type": "object",
				"additionalProperties": false
			},
			"RequestUserGetUsername": {
				"properties": {
					"user_uuid": {
						"type": "string"
					},
					"session_hash": {
						"type": "string"
					},
					"request_hash": {
						"type": "string"
					},
					"timestamp": {
						"type": "number",
						"format": "double"
					},
					"requested_user_uuid": {
						"type": "string"
					}
				},
				"required": [
					"user_uuid",
					"session_hash",
					"request_hash",
					"timestamp",
					"requested_user_uuid"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "ts-task-backend",
		"version": "1.0.0",
		"license": {
			"name": "MIT"
		},
		"contact": {
			"name": "pukanjourpour"
		}
	},
	"openapi": "3.0.0",
	"paths": {
...
		"/users/register": {
			"post": {
				"operationId": "Register",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ResponseUserRegister"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/RequestUserRegister"
							}
						}
					}
				}
			}
		},
		"/users/login": {
			"post": {
				"operationId": "Login",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ResponseUserLogin"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/RequestUserLogin"
							}
						}
					}
				}
			}
		},
		"/users/get-username": {
			"post": {
				"operationId": "GetUsernameByUuid",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ResponseUserGetUsername"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/RequestUserGetUsername"
							}
						}
					}
				}
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}